Alpha-quality temporal safety disk image and kernels.
These kernels require an updated FPGA image or QEMU with CSetEqualExact.

Updating QEMU may require an updated BBL.

We're tracking a bug where non-sensible capabilities are being presented
for revocation in the kernel on FPGA. By disabling some optimizations we
have a kernel that proceeds with revocation in this case.